+++ src/Backend.php
@@ -1,0 +1,44 @@
+<?php
+/**
+ * @brief checkStoreVersion, a plugin for Dotclear 2
+ *
+ * @package Dotclear
+ * @subpackage Plugin
+ *
+ * @author Jean-Christian Denis and Contributors
+ *
+ * @copyright Jean-Christian Denis
+ * @copyright GPL-2.0 https://www.gnu.org/licenses/gpl-2.0.html
+ */
+declare(strict_types=1);
+
+namespace Dotclear\Plugin\checkStoreVersion;
+
+use dcCore;
+use dcNsProcess;
+
+class Backend extends dcNsProcess
+{
+    public static function init(): bool
+    {
+        if (defined('DC_CONTEXT_ADMIN')) {
+            self::$init = dcCore::app()->auth->isSuperAdmin();
+        }
+
+        return self::$init;
+    }
+
+    public static function process(): bool
+    {
+        if (!self::$init) {
+            return false;
+        }
+
+        dcCore::app()->addBehaviors([
+            'pluginsToolsTabsV2' => [BackendBehaviors::class, 'pluginTabs'],
+            'themesToolsTabsV2'  => [BackendBehaviors::class, 'themesTabs'],
+        ]);
+
+        return true;
+    }
+}