+++ src/My.php
@@ -1,0 +1,20 @@
+<?php
+/**
+ * @brief checkStoreVersion, a plugin for Dotclear 2
+ *
+ * @package Dotclear
+ * @subpackage Plugin
+ *
+ * @author Jean-Christian Denis and Contributors
+ *
+ * @copyright Jean-Christian Denis
+ * @copyright GPL-2.0 https://www.gnu.org/licenses/gpl-2.0.html
+ */
+declare(strict_types=1);
+
+namespace Dotclear\Plugin\checkStoreVersion;
+
+class My
+{
+    public const DC_MAX = '2.25.3';
+}