+++ CHANGELOG.md
@@ -1,4 +1,6 @@
-x.x
+0.3 - 2023.03.11
+- required Dotclear 2.25, up to 2.25.3
+- use php namespace
 - fix translation
 
 0.2 - 2022.12.03
@@ -6,5 +8,5 @@ x.x
 - Add Dotclear maximum version compatibility (thanks @Gvx- )
 
 0.1.1 - 2022.12.03
-- Required Dotclear 2.19
+- Required Dotclear 2.19, up to 2.23.1
 - First release